refactor(guest-mock-codex): split integration tests
